Total restyling
HTML Modifications
--------------------------------------------------
I chose a background image for the page, since I
have not trained graphic intuition yet. The chosen
style is minimal: a main div, wrapping the other elements.
The website is not responsive yet.
--------------------------------------------------

CSS Modifications
--------------------------------------------------
Minor modificatons, reflected, mainly, on the
arrangement of the elements in the HTML page.
--------------------------------------------------

JS Modifications
--------------------------------------------------
The logic of the game pretty much stayed the same,
with some minor modifications. The only thing which
stands out in respect to the previous version is
the massive use of event listeners. Most of the
lines of this code are abundant, when they could
be written in a short way, I am aware of this.

In fact, in the future, I plan to rewrite the code
with the use of regular expressions.
--------------------------------------------------